feat(renderer): bootstrap React root behind rollback gate (#74)
