+++ package.json
@@ -39,10 +39,13 @@
   "devDependencies": {
     "@playwright/test": "^1.58.2",
     "@types/node": "22.10.2",
+    "@types/react": "18.3.12",
+    "@types/react-dom": "18.3.1",
     "@vitest/coverage-v8": "2.1.8",
     "electron": "38.0.0",
     "electron-builder": "26.0.12",
     "electron-vite": "4.0.0",
+    "jsdom": "24.1.3",
     "playwright": "^1.58.2",
     "typescript": "5.7.3",
     "vite": "6.0.7",
@@ -50,6 +53,8 @@
   },
   "dependencies": {
     "electron-store": "^11.0.2",
+    "react": "18.3.1",
+    "react-dom": "18.3.1",
     "valibot": "^1.2.0"
   },
   "pnpm": {
